add licensing
Adding License, because multiple people want to push updates for this, but are forced to ask me without licesensing; which is now no longer the case. 
This is open source and you're free to do whatever you wish with the code as long as you credit this repo as master. 👍